LayoutTests/imported/w3c/web-platform-tests/html/anonymous-iframe/worker-cookies.tentative.https.window-expected.txt


PASS set cookies
PASS Worker spawned from normal iframe can access global cookies
FAIL Worker spawned from credentialless iframe can't access global cookies assert_equals: expected (undefined) undefined but got (string) "cookie_value"
